requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
autopep8==2.0.1
black==22.12.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
docker-pycreds==0.4.0
filelock==3.8.2
fire==0.4.0
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.29
huggingface-hub
idna==3.4
joblib==1.2.0
lightning-utilities==0.3.0
multidict==6.0.3
mypy-extensions==0.4.3
nltk==3.8.1
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pathspec==0.10.3
pathtools==0.1.2
Pillow==9.2.0
pip==22.3.1
platformdirs==2.6.2
promise==2.3
protobuf==3.20.1
psutil==5.9.4
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.8.3.post1
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
sentry-sdk==1.11.1
setproctitle==1.3.2
setuptools==65.5.1
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
tensorboardX==2.5.1
termcolor==2.1.1
tokenizers==0.13.2
# torch==1.10.1
# torchaudio==0.10.1
# torchmetrics==0.11.0
# torchvision==0.11.2
tqdm==4.64.1
transformers==4.31.0
typing_extensions==4.4.0
urllib3==1.26.13
wandb==0.13.6
wheel==0.38.4
yarl==1.8.2
scikit-learn==1.2.0
seaborn==0.12.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
ipython==8.9.0
nltk==3.8.1
peft==0.4.0
hydra-core==1.3.0
sentencepiece
deepspeed